feat(restaurantDBService): 음식점 DB 서비스 도메인 구현
